requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
PyJWT==2.8.0
bcrypt==4.0.1
pyotp==2.9.0
qrcode==7.4
Pillow==10.0.1
python-dotenv==1.0.0
